+++ folib-web-core/src/main/java/com/veadan/folib/gremlin/common/ArtifactsResult.java
@@ -3,6 +3,8 @@ package com.veadan.folib.gremlin.common;
 import lombok.Data;
 import lombok.experimental.Accessors;
 
+import java.math.BigDecimal;
+
 @Data
 @Accessors(chain = true)
 public class ArtifactsResult {
@@ -11,7 +13,7 @@ public class ArtifactsResult {
     private String repositoryId;
     private String artifactName;
     private String artifactPath;
-    private Long sizeInBytes;
+    private BigDecimal sizeInBytes;
     private String path;
     
 }
